# src/docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Muabantainguyen API",
		"version": "0.2.0",
		"description": "API documentation with admin management, permissions, and site settings"
	},
	"servers": [{ "url": "http://localhost:4000" }],
	"paths": {
		"/api/admin/admins": {
			"get": {
				"summary": "List all admins with roles and permissions",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Management"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean" },
										"message": { "type": "string" },
										"admins": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": { "type": "string", "format": "uuid" },
													"name": { "type": "string" },
													"email": { "type": "string" },
													"phone": { "type": "string" },
													"role": { "type": "string", "enum": ["admin", "super", "staff"] },
													"admin_role_name": { "type": "string" },
													"admin_role_display": { "type": "string" },
													"permissions": { "type": "array", "items": { "type": "string" } },
													"is_blocked": { "type": "boolean" },
													"created_at": { "type": "string", "format": "date-time" }
												}
											}
										}
									}
								}
							}
						}
					},
					"403": { "description": "Forbidden - requires admins.view permission" }
				}
			},
			"post": {
				"summary": "Create new admin account",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Management"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": { "type": "string", "example": "John Admin" },
									"email": { "type": "string", "format": "email", "example": "admin@shop.com" },
									"phone": { "type": "string", "example": "0123456789" },
									"password": { "type": "string", "example": "securepassword123" },
									"adminRoleId": { "type": "string", "format": "uuid", "example": "uuid-of-role" }
								},
								"required": ["email", "password", "adminRoleId"]
							}
						}
					}
				},
				"responses": {
					"201": { "description": "Admin created successfully" },
					"400": { "description": "Bad request - missing required fields" },
					"403": { "description": "Forbidden - requires admins.create permission" }
				}
			}
		},
		"/api/admin/admins/{id}/role": {
			"put": {
				"summary": "Update admin role",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Management"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"adminRoleId": { "type": "string", "format": "uuid" }
								},
								"required": ["adminRoleId"]
							}
						}
					}
				},
				"responses": {
					"200": { "description": "Admin role updated successfully" },
					"403": { "description": "Forbidden - requires admins.edit permission" }
				}
			}
		},
		"/api/admin/roles": {
			"get": {
				"summary": "List all admin roles with permissions",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Roles & Permissions"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean" },
										"message": { "type": "string" },
										"roles": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": { "type": "string", "format": "uuid" },
													"name": { "type": "string", "example": "product_manager" },
													"display_name": { "type": "string", "example": "Product Manager" },
													"description": { "type": "string" },
													"is_active": { "type": "boolean" },
													"permissions": {
														"type": "array",
														"items": {
															"type": "object",
															"properties": {
																"name": {
																	"type": "string",
																	"example": "products.view"
																},
																"display_name": {
																	"type": "string",
																	"example": "Xem sản phẩm"
																},
																"module": { "type": "string", "example": "products" }
															}
														}
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"403": { "description": "Forbidden - requires admins.view permission" }
				}
			}
		},
		"/api/admin/permissions": {
			"get": {
				"summary": "List all permissions grouped by module",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Roles & Permissions"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean" },
										"message": { "type": "string" },
										"permissions": {
											"type": "object",
											"additionalProperties": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"id": { "type": "string", "format": "uuid" },
														"name": { "type": "string", "example": "products.create" },
														"display_name": { "type": "string", "example": "Tạo sản phẩm" },
														"description": { "type": "string" },
														"module": { "type": "string", "example": "products" }
													}
												}
											},
											"example": {
												"products": [
													{
														"name": "products.view",
														"display_name": "Xem sản phẩm",
														"module": "products"
													}
												],
												"orders": [
													{
														"name": "orders.view",
														"display_name": "Xem đơn hàng",
														"module": "orders"
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"403": { "description": "Forbidden - requires admins.view permission" }
				}
			}
		},
		"/api/admin/settings": {
			"get": {
				"summary": "Get all site settings (admin only)",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Site Settings"],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean" },
										"message": { "type": "string" },
										"settings": {
											"type": "object",
											"additionalProperties": {
												"type": "object",
												"additionalProperties": {
													"type": "object",
													"properties": {
														"value": { "type": "string" },
														"type": {
															"type": "string",
															"enum": ["text", "number", "boolean", "json", "image"]
														},
														"display_name": { "type": "string" },
														"description": { "type": "string" },
														"is_public": { "type": "boolean" }
													}
												}
											},
											"example": {
												"branding": {
													"site_name": {
														"value": "My Shop",
														"type": "text",
														"display_name": "Tên website",
														"is_public": true
													},
													"site_logo": {
														"value": "https://logo-url.com",
														"type": "image",
														"display_name": "Logo website",
														"is_public": true
													}
												},
												"contact": {
													"contact_email": {
														"value": "contact@shop.com",
														"type": "text",
														"display_name": "Email liên hệ",
														"is_public": true
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"403": { "description": "Forbidden - requires settings.view permission" }
				}
			},
			"put": {
				"summary": "Update site settings",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Site Settings"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": { "type": "string" },
								"example": {
									"site_name": "My New Shop Name",
									"site_logo": "https://new-logo-url.com",
									"contact_email": "newcontact@shop.com",
									"maintenance_mode": "false",
									"allow_registration": "true"
								}
							}
						}
					}
				},
				"responses": {
					"200": { "description": "Settings updated successfully" },
					"403": { "description": "Forbidden - requires settings.edit permission" }
				}
			}
		},
...
		"/api/admin/products/{id}": {
			"get": {
				"summary": "Admin: get product detail",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Products"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": { "type": "boolean" },
										"message": { "type": "string" },
										"product": {
											"type": "object",
											"properties": {
												"id": { "type": "string", "format": "uuid" },
												"name": { "type": "string" },
												"slug": { "type": "string" },
												"description": { "type": "string" },
												"price": { "type": "number" },
												"stock": { "type": "integer" },
												"image_url": { "type": "string" },
												"category_id": { "type": "string", "format": "uuid" },
												"category_name": { "type": "string" },
												"category_slug": { "type": "string" },
												"is_active": { "type": "boolean" },
												"created_at": { "type": "string", "format": "date-time" },
												"updated_at": { "type": "string", "format": "date-time" }
											}
										}
									}
								}
							}
						}
					},
					"404": { "description": "Product not found" },
					"403": { "description": "Forbidden - requires products.view permission" }
				}
			},
			"put": {
				"summary": "Admin: update product",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Products"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": { "type": "string" },
									"slug": { "type": "string" },
									"description": { "type": "string" },
									"price": { "type": "number" },
									"stock": { "type": "integer" },
									"imageUrl": { "type": "string" },
									"categoryId": { "type": "string", "format": "uuid" },
									"isActive": { "type": "boolean" }
								}
							}
						}
					}
				},
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires products.edit permission" }
				}
			},
			"delete": {
				"summary": "Admin: delete product",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Products"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires products.delete permission" }
				}
			}
		},
		"/api/admin/products": {
			"post": {
				"summary": "Admin: create product",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Products"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": { "type": "string" },
									"slug": { "type": "string" },
									"description": { "type": "string" },
									"price": { "type": "number" },
									"stock": { "type": "integer" },
									"imageUrl": { "type": "string" },
									"categoryId": { "type": "string", "format": "uuid" }
								},
								"required": ["name", "slug", "price"]
							}
						}
					}
				},
				"responses": {
					"201": { "description": "Created" },
					"403": { "description": "Forbidden - requires products.create permission" }
				}
			}
		},
		"/api/admin/categories": {
			"post": {
				"summary": "Admin: create category",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Categories"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": { "type": "string" },
									"slug": { "type": "string" },
									"parentId": { "type": "string", "format": "uuid" }
								},
								"required": ["name", "slug"]
							}
						}
					}
				},
				"responses": {
					"201": { "description": "Created" },
					"403": { "description": "Forbidden - requires categories.manage permission" }
				}
			}
		},
		"/api/admin/categories/{id}": {
			"put": {
				"summary": "Admin: update category",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Categories"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires categories.manage permission" }
				}
			},
			"delete": {
				"summary": "Admin: delete category",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Categories"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires categories.manage permission" }
				}
			}
		},
		"/api/admin/categories/tree": {
			"get": {
				"summary": "Admin: get category tree",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Categories"],
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires categories.manage permission" }
				}
			}
		},
		"/api/admin/users": {
			"get": {
				"summary": "Admin: list users with pagination, search and filters",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Users"],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": { "type": "integer", "default": 1 },
						"description": "Page number"
					},
					{
						"name": "pageSize",
						"in": "query",
						"schema": { "type": "integer", "default": 20 },
						"description": "Number of items per page"
					},
					{
						"name": "search",
						"in": "query",
						"schema": { "type": "string" },
						"description": "Search in name, email, phone"
					},
					{
						"name": "role",
						"in": "query",
						"schema": { "type": "string", "enum": ["user", "admin", "staff"] },
						"description": "Filter by role"
					},
					{
						"name": "isBlocked",
						"in": "query",
						"schema": { "type": "string", "enum": ["true", "false", ""] },
						"description": "Filter by blocked status"
					},
					{
						"name": "sortBy",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["created_at", "name", "email", "role"],
							"default": "created_at"
						},
						"description": "Sort field"
					},
					{
						"name": "sortOrder",
						"in": "query",
						"schema": { "type": "string", "enum": ["ASC", "DESC"], "default": "DESC" },
						"description": "Sort order"
					}
				],
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires users.view permission" }
				}
			}
		},
		"/api/admin/users/{id}/block": {
			"put": {
				"summary": "Admin: block/unblock user",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Users"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": { "blocked": { "type": "boolean" } },
								"required": ["blocked"]
							}
						}
					}
				},
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires users.block permission" }
				}
			}
		},
		"/api/admin/users/{id}/orders": {
			"get": {
				"summary": "Admin: get user orders",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Users"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": {
					"200": { "description": "OK" },
					"403": { "description": "Forbidden - requires orders.view permission" }
				}
			}
		},
		"/api/admin/vouchers": {
			"get": {
				"summary": "Admin: list vouchers",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Vouchers"],
				"responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" } }
			},
			"post": {
				"summary": "Admin: create voucher",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Vouchers"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"code": { "type": "string" },
									"description": { "type": "string" },
									"discount_percent": { "type": "integer" },
									"discount_amount": { "type": "number" },
									"max_uses": { "type": "integer" },
									"valid_from": { "type": "string", "format": "date-time" },
									"valid_to": { "type": "string", "format": "date-time" },
									"is_active": { "type": "boolean" }
								},
								"required": ["code"]
							}
						}
					}
				},
				"responses": { "201": { "description": "Created" }, "403": { "description": "Forbidden" } }
			}
		},
		"/api/admin/vouchers/{id}": {
			"put": {
				"summary": "Admin: update voucher",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Vouchers"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" } }
			},
			"delete": {
				"summary": "Admin: delete voucher",
				"security": [{ "bearerAuth": [] }],
				"tags": ["Admin Vouchers"],
				"parameters": [
					{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
				],
				"responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" } }
			}
		},
...
		"/api/products": {
			"get": {
				"summary": "List products with filters",
				"tags": ["Products"],
				"parameters": [
					{ "name": "q", "in": "query", "schema": { "type": "string" } },
					{ "name": "categoryId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
					{ "name": "minPrice", "in": "query", "schema": { "type": "number" } },
					{ "name": "maxPrice", "in": "query", "schema": { "type": "number" } },
					{ "name": "inStock", "in": "query", "schema": { "type": "boolean" } },
					{ "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
					{ "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } }
				],
				"responses": { "200": { "description": "OK" } }
			}
		},
		"/api/products/{slug}": {
			"get": {
				"summary": "Get product detail by slug",
				"tags": ["Products"],
				"parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }],
				"responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
...
	},
...
}